% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRaceList.R
\name{get_race_list}
\alias{get_race_list}
\title{Get Race List}
\usage{
get_race_list(year = 2022, series = 1)
}
\arguments{
\item{series}{1=Cup, 2=Xfinity, 3=Trucks.}

\item{years}{The year 2022.}
}
\value{
A data frame of race information.
}
\description{
Pulls Race List for seasons and series mentioned. This dataset contains the race ids.
}
\examples{
get_race_list(year=2022,series=2)
get_race_list(year=c(2021,2022),series=c(1,2,3))
}
