% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSchedule.R
\name{get_schedule}
\alias{get_schedule}
\title{Get Schedule}
\usage{
get_schedule(years = 2022)
}
\arguments{
\item{years}{The year 2022.}
}
\value{
A data frame of schedule information.
}
\description{
Pulls Weekend Schedule for seasons mentioned.
}
\examples{
get_schedule(year=2022)
get_schedule(year=c(2021,2022))
}
