% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSpeeds.R
\name{get_speeds}
\alias{get_speeds}
\title{Get Speeds}
\usage{
get_speeds(year = 2022, series = 1)
}
\arguments{
\item{year}{The year 2022.}

\item{series}{1=Cup, 2=Xfinity, 3=Trucks.}
}
\value{
A data frame of laps speeds.
}
\description{
Gets Individual Lap Speed Data for the 2022 NASCAR Season
}
\examples{
get_speeds(year=2022,series=2)
}
