% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nas_lap_times.R
\name{load_nas_lap_times}
\alias{load_nas_lap_times}
\title{Load NAS Lap Time}
\usage{
load_nas_lap_times(rid, time_pause = 1)
}
\arguments{
\item{rid}{Race ID}

\item{time_pause}{Wait time (in seconds) between page loads}
}
\value{
A data frame containing lap time data
}
\description{
Loads Individual Lap Time Data for any race after 2020
}
\examples{
load_nas_lap_times(rid=5009)

}
