% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nas_live_points.R
\name{load_nas_live_points}
\alias{load_nas_live_points}
\title{Load Nas Live Points}
\usage{
load_nas_live_points(year, time_pause = 1)
}
\arguments{
\item{year}{2022}

\item{time_pause}{Wait time (in seconds) between page loads}
}
\value{
A data frame containing lap time data
}
\description{
Gets Live Points for any season after 2019
}
\examples{
load_nas_live_points(year=2022)

}
