% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nas_pit_stops.R
\name{load_nas_pit_stops}
\alias{load_nas_pit_stops}
\title{Load Nas Pit Stops}
\usage{
load_nas_pit_stops(year, time_pause = 1)
}
\arguments{
\item{year}{2022}

\item{time_pause}{Wait time (in seconds) between page loads}
}
\value{
A data frame containing pit stop data
}
\description{
Gets pit stop data for any season after 2020
}
\examples{
load_nas_pit_stops(year=2022)

}
