% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nas_race_cautions.R
\name{load_nas_race_cautions}
\alias{load_nas_race_cautions}
\title{Load Nas Race Cautions}
\usage{
load_nas_race_cautions(year, time_pause = 1)
}
\arguments{
\item{year}{2022}

\item{time_pause}{Wait time (in seconds) between page loads}
}
\value{
A data frame containing race cautions data
}
\description{
Gets Load Race Cautions for any season after 2018
}
\examples{
load_nas_race_cautions(year=2022)

}
