% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nas_race_info.R
\name{load_nas_race_info}
\alias{load_nas_race_info}
\title{Load Nas Race Info}
\usage{
load_nas_race_info(year, time_pause = 1)
}
\arguments{
\item{year}{2022}

\item{time_pause}{Wait time (in seconds) between page loads}
}
\value{
A data frame containing race info data
}
\description{
Gets Race Info for any season after 2018
}
\examples{
load_nas_race_info(year=2022)

}
