% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nas_race_infractions.R
\name{load_nas_race_infractions}
\alias{load_nas_race_infractions}
\title{Load Nas Race Infractions}
\usage{
load_nas_race_infractions(year, time_pause = 1)
}
\arguments{
\item{year}{2022}

\item{time_pause}{Wait time (in seconds) between page loads}
}
\value{
A data frame containing race infractions data
}
\description{
Gets Race Infractions for any season after 2018
}
\examples{
load_nas_race_infractions(year=2022)

}
