% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nas_race_leaders.R
\name{load_nas_race_leaders}
\alias{load_nas_race_leaders}
\title{Load Nas Race Leaders}
\usage{
load_nas_race_leaders(year, time_pause = 1)
}
\arguments{
\item{year}{2022}

\item{time_pause}{Wait time (in seconds) between page loads}
}
\value{
A data frame containing Race Leaders data
}
\description{
Gets Race Leaders for any season after 2018
}
\examples{
load_nas_race_leaders(year=2022)

}
