% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nas_race_results.R
\name{load_nas_race_results}
\alias{load_nas_race_results}
\title{Load Nas Race Results}
\usage{
load_nas_race_results(year, time_pause = 1)
}
\arguments{
\item{year}{2022}

\item{time_pause}{Wait time (in seconds) between page loads}
}
\value{
A data frame containing race results data
}
\description{
Gets race results data for any season after 2018
}
\examples{
load_nas_race_results(year=2022)

}
