% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nas_schedule.R
\name{load_nas_schedule}
\alias{load_nas_schedule}
\title{Load Nas Schedule}
\usage{
load_nas_schedule(year, time_pause = 1)
}
\arguments{
\item{year}{2022}

\item{time_pause}{Wait time (in seconds) between page loads}
}
\value{
A data frame containing schedule data
}
\description{
Gets Schedule results for any season after 2019
}
\examples{
load_nas_schedule(year=2022)

}
