% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nas_weekend.R
\name{load_nas_weekend}
\alias{load_nas_weekend}
\title{Load Nas Weekend Results}
\usage{
load_nas_weekend(year, time_pause = 1)
}
\arguments{
\item{year}{2022}

\item{time_pause}{Wait time (in seconds) between page loads}
}
\value{
A data frame containing weekend data
}
\description{
Gets weekend results for any season after 2015
}
\examples{
load_nas_weekend(year=2022)

}
