% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rr_schedule.R
\name{load_rr_schedule}
\alias{load_rr_schedule}
\title{Load RR Schedule}
\usage{
load_rr_schedule(year, series, time_pause = 1)
}
\arguments{
\item{year}{numeric year with format YYYY}

\item{series}{cup, xfinity, trucks, arca, indycar, ...}

\item{time_pause}{Wait time (in seconds) between page loads}
}
\value{
data.frame of the racing-reference schedule
}
\description{
Loads the Racing Reference Schedule for the following series:
a1_grand_prix      | A1 Grand Prix
act_late_model     | ACT Late Model
asa_national_tour  | ASA National Tour
cars_tour          | CARS Tour
cart               | CART
indy_lights        | Indy NXT
indycar            | Indycar
iroc               | IROC
nascar_arca        | NASCAR ARCA main series
nascar_arca_east   | NASCAR ARCA east
nascar_arca_west   | NASCAR ARCA west
nascar_busch       | NASCAR Xfinity series
nascar_convertible | NASCAR Convertible Division
nascar_cup         | NASCAR Cup series
nascar_fedex_challenge     | NASCAR FedEx Challenge
nascar_grand_national_east | NASCAR Grand National East
nascar_midwest_series      | NASCAR Midwest Series
nascar_modified    | NASCAR Whelen Modified Tour
nascar_north_tour  | NASCAR North Tour
nascar_northwest_series    | NASCAR Northwest Series
nascar_peak_mexico | NASCAR Peak Mexico Series
nascar_pintys      | NASCAR Pinty's Series
nascar_southeast_series    | NASCAR Southeast Series
nascar_southwest_series    | NASCAR Southwest Series
nascar_trucks      | NASCAR Truck Series
nascar_whelen_southern_modified | NASCAR Whelen Southern Modified Tour
natcc              | North American Touring Car Championship
}
\examples{
load_rr_schedule(year=2022, series="cup")

}
