% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas_lap_times.R
\name{nas_lap_times}
\alias{nas_lap_times}
\title{Nas Lap Time}
\usage{
nas_lap_times(rid)
}
\arguments{
\item{rid}{Race ID}
}
\value{
A data frame containing lap time data
}
\description{
Gets Individual Lap Time Data for any race after 2020
}
\examples{
nas_lap_times(rid=5009)

}
