% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas_live_points.R
\name{nas_live_points}
\alias{nas_live_points}
\title{Nas Live Points}
\usage{
nas_live_points(year)
}
\arguments{
\item{year}{2022}
}
\value{
A data frame containing lap time data
}
\description{
Gets Live Points for any season after 2019
}
\examples{
nas_live_points(year=2022)

}
