% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas_pit_stops.R
\name{nas_pit_stops}
\alias{nas_pit_stops}
\title{Nas Pit Stops}
\usage{
nas_pit_stops(year)
}
\arguments{
\item{year}{2022}
}
\value{
A data frame containing pit stop data
}
\description{
Gets pit stop data for any season after 2020
}
\examples{
nas_pit_stops(year=2022)

}
