% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas_pits.R
\name{nas_pits}
\alias{nas_pits}
\title{Get Pit Data}
\usage{
nas_pits(year = 2022, series = 1)
}
\arguments{
\item{year}{The year 2022.}

\item{series}{1=Cup, 2=Xfinity, 3=Trucks.}
}
\value{
A data frame of pit stops.
}
\description{
Gets Pit Stop Data for the 2022 NASCAR Season from NASCAR.com
}
\examples{
nas_pits(year=2022,series=2)
}
