% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas_practice.R
\name{nas_practice}
\alias{nas_practice}
\title{Nascar Practice Data}
\usage{
nas_practice(year = 2022, series = 1)
}
\arguments{
\item{series}{1=Cup, 2=Xfinity, 3=Trucks.}

\item{years}{The year 2022.}
}
\value{
A data frame of race information.
}
\description{
Pulls Practice Lap Averages for seasons and series mentioned from NASCAR.com. This dataset contains the race ids.
}
\examples{
nas_practice(year=2022,series=2)
nas_practice(year=c(2021,2022),series=c(1,2,3))

}
