% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas_race_cautions.R
\name{nas_race_cautions}
\alias{nas_race_cautions}
\title{Nas Race Cautions}
\usage{
nas_race_cautions(year)
}
\arguments{
\item{year}{2022}
}
\value{
A data frame containing race cautions data
}
\description{
Gets Race Cautions for any season after 2018
}
\examples{
nas_race_cautions(year=2022)

}
