% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas_race_info.R
\name{nas_race_info}
\alias{nas_race_info}
\title{Nas Race Info}
\usage{
nas_race_info(year)
}
\arguments{
\item{year}{2022}
}
\value{
A data frame containing race info data
}
\description{
Gets Race Info for any season after 2018
}
\examples{
nas_race_info(year=2022)

}
