% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas_race_infractions.R
\name{nas_race_infractions}
\alias{nas_race_infractions}
\title{Nas Race Infractions}
\usage{
nas_race_infractions(year)
}
\arguments{
\item{year}{2022}
}
\value{
A data frame containing race infractions data
}
\description{
Gets Race Infractions for any season after 2018
}
\examples{
nas_race_infractions(year=2022)

}
