% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas_race_leaders.R
\name{nas_race_leaders}
\alias{nas_race_leaders}
\title{Nas Race Leaders}
\usage{
nas_race_leaders(year)
}
\arguments{
\item{year}{2022}
}
\value{
A data frame containing Race Leaders data
}
\description{
Gets Race Leaders for any season after 2018
}
\examples{
nas_race_leaders(year=2022)

}
