% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas_race_list.R
\name{nas_race_list}
\alias{nas_race_list}
\title{Nascar Race List}
\usage{
nas_race_list(year = 2022, series = 1)
}
\arguments{
\item{series}{1=Cup, 2=Xfinity, 3=Trucks.}

\item{years}{The year 2022.}
}
\value{
A data frame of race information.
}
\description{
Pulls Race List for seasons and series mentioned from NASCAR.com. This dataset contains the race ids.
}
\examples{
nas_race_list(year=2022,series=2)
nas_race_list(year=c(2021,2022),series=c(1,2,3))
}
