% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas_race_results.R
\name{nas_race_results}
\alias{nas_race_results}
\title{Nas Race Results}
\usage{
nas_race_results(year)
}
\arguments{
\item{year}{2022}
}
\value{
A data frame containing race results data
}
\description{
Gets race results data for any season after 2018
}
\examples{
nas_race_results(year=2022)

}
