% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas_race_stage_results.R
\name{nas_race_stage_results}
\alias{nas_race_stage_results}
\title{Nas Stage Race Results}
\usage{
nas_race_stage_results(year)
}
\arguments{
\item{year}{2022}
}
\value{
A data frame containing race stage results data
}
\description{
Gets Stage Race results for any season after 2019
}
\examples{
nas_race_stage_results(year=2022)

}
