% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas_schedule.R
\name{nas_schedule}
\alias{nas_schedule}
\title{NAS Schedule}
\usage{
nas_schedule(years = 2022)
}
\arguments{
\item{years}{The year 2022.}
}
\value{
A data frame of schedule information.
}
\description{
Pulls Weekend Schedule for seasons mentioned.
}
\examples{
nas_schedule(year=2022)
nas_schedule(year=c(2021,2022))
}
