% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas_speeds.R
\name{nas_speeds}
\alias{nas_speeds}
\title{Nas Speeds}
\usage{
nas_speeds(year = 2022, series = 1)
}
\arguments{
\item{year}{The year 2022.}

\item{series}{1=Cup, 2=Xfinity, 3=Trucks.}
}
\value{
A list of data frames containing lap speeds.
}
\description{
Gets Individual Lap Speed Data for the 2022 NASCAR Season
}
\examples{
nas_speeds(year=2022,series=2)

}
