% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas_weekend.R
\name{nas_weekend}
\alias{nas_weekend}
\title{Nas Weekend Results}
\usage{
nas_weekend(year)
}
\arguments{
\item{year}{2022}
}
\value{
A data frame containing weekend data
}
\description{
Gets weekend results for any season after 2015
}
\examples{
nas_weekend(year=2022)

}
