% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_loopdata.R
\name{rr_loopdata}
\alias{rr_loopdata}
\title{Pull Loop Data from racing-reference.info}
\usage{
rr_loopdata(races = "all", year = 2022, series = 1)
}
\arguments{
\item{races}{Either in race_id format (5146) or rr_id format ("01"), defaults to "all" races}

\item{series}{1=Cup, 2=Xfinity, 3=Trucks.}

\item{years}{A single year 2015-2022.}
}
\value{
Loop data for the specified races.
}
\description{
Pulls Loop Data for the 2022 NASCAR Season
}
\examples{
rr_loopdata(races=5146, year=2022,series=2)
rr_loopdata(races="01", year=2022,series=2)
rr_loopdata(races="all", year=2022,series=2)
}
