% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_series_name.R
\name{standardize_series_name}
\alias{standardize_series_name}
\title{Standardize Series Name}
\usage{
standardize_series_name(series)
}
\arguments{
\item{series}{cup, xfinity, trucks, arca, indycar, ...}
}
\value{
series_id
}
\description{
Inputs series identifier and returns standardized series ID
}
\examples{
standardize_series_name(series="MONSTER cup sEriES")

}
