% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwalk_race_ids.R
\name{xwalk_race_ids}
\alias{xwalk_race_ids}
\title{Creates a Crosswalk for race_ids from racing-reference.info and NASCAR.com}
\usage{
xwalk_race_ids()
}
\value{
data.frame
}
\description{
Creates a Crosswalk for race_ids from racing-reference.info and NASCAR.com. Contains all races 2015-2023
}
\examples{
xwalk_race_ids()
rr_loopdata(races=1, year=2022,series=2)
}
