% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas_schedule.R
\name{nas_schedule}
\alias{nas_schedule}
\title{Nas Schedule}
\usage{
nas_schedule(year)
}
\arguments{
\item{year}{2022}
}
\value{
A data frame containing schedule data
}
\description{
Gets Schedule results for any season after 2019
}
\examples{
nas_schedule(year=2022)

}
